+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Translation of `pandas.MultiIndex` is now properly supported, as indicated by `resolve_io` not throwing.
+
 ## [0.4.0] - 2023-06-16
 
 ### Added